Jetpack Licensing: Rename types.ts to index.d.ts to match TS convention
