+++ client/state/licensing-portal/selectors/inspect-license.ts
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { LicensingPortalStore } from 'calypso/state/licensing-portal/types';
+import { LicensingPortalStore } from 'calypso/state/licensing-portal';
 // Required for modular state.
 import 'calypso/state/licensing-portal/init';
